Move User service to `service`
